Fixed EZPNEXT-679: Connections to memcached aren't closed
Short: Memcached:__construct() implements the Multiton Pattern.

Longer: When using persistent memcached connections, those need to be
initialized once, the extension take care of returning the zval that has
been previously initialized with the same key. It means that it already
has the connection to the servers as well as the options.